chore: update eframe version
Updates eframe to v0.32.0

Updates the eframe and egui_extras dependencies to version 0.32.0.

This brings in the latest features, improvements, and bug fixes from the eframe ecosystem.

Also includes minor code adjustments to align with the updated eframe API, such as use of UiBuilder and StrokeKind enum, and the removal of clamp_to_range